requirements.txt
annotated-types==0.7.0
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cymem==2.0.8
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
preshed==3.0.9
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pymongo==4.7.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.14.0
selenium==4.21.0
setuptools==70.1.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.5
threadpoolctl==3.5.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
